pyproject.toml
[tool.poetry]
name = "lightrag"
version = "0.1.0"
description = "1000 lines of code is all you need. Be light and be lightening fast."
authors = ["Li Yin <li.yin.gravity@gmail.com>"]
readme = "README.md"
include = ["lightrag", "extend", "tests", "use_cases"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0"
backoff = "^2.2.1"
openai = "^1.12.0"
jinja2 = "^3.1.3"
pytest = "^8.1.1"
groq = "^0.5.0"

[tool.poetry.group.dev.dependencies]
langchain = "^0.1.16"
langchainhub = "^0.1.15"
llama-index = "^0.10.30"
pre-commit = "^3.7.0"
torch = "^2.3.0"
dspy-ai = "^2.4.9"
litellm = "^1.35.34"
haystack-ai = "^2.0.1"
transformers = "^4.40.1"
pgvector-haystack = "^0.1.0"
ollama = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
